+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features added:
 * Cover `<MatchingFieldsManager>` component with tests (UIDATIMP-712)
 * Cover `<EditKeyShortcutsWrapper` component with tests (UIDATIMP-957)
+* Cover `<ReturnToAssignJobs>` component with tests (UIDATIMP-960)
 
 ## [5.0.0](https://github.com/folio-org/ui-data-import/tree/v5.0.0) (2021-10-08)